agent/requirements.txt
strands-agents>=1.30.0
strands-agents-tools>=0.2.0
bedrock-agentcore>=1.4.0
boto3>=1.42.0
